Add document for examples
