+++ .bumpversion.toml
@@ -1,0 +1,9 @@
+# SPDX-FileCopyrightText: 2024 Shun Sakai
+#
+# SPDX-License-Identifier: Apache-2.0 OR MIT
+
+[tool.bumpversion]
+current_version = "0.2.0"
+
+[[tool.bumpversion.files]]
+filename = "examples/examples.go"
