--- examples/README.adoc
@@ -1,26 +1,0 @@
-// SPDX-FileCopyrightText: 2024 Shun Sakai
-//
-// SPDX-License-Identifier: Apache-2.0 OR MIT
-
-= Examples
-
-link:encrypt/main.go[]::
-
-  An example of encrypting to the abcrypt encrypted data format.
-
-link:decrypt/main.go[]::
-
-  An example of decrypting from the abcrypt encrypted data format.
-
-link:info/main.go[]::
-
-  An example of reading the Argon2 parameters.
-
-== How to build the example
-
-To build these programs run the following in the project root directory.
-
-[source,sh]
-----
-go build -o . ./...
-----
