+++ examples/README.md
@@ -1,0 +1,29 @@
+<!--
+SPDX-FileCopyrightText: 2024 Shun Sakai
+
+SPDX-License-Identifier: Apache-2.0 OR MIT
+-->
+
+# Examples
+
+## Programs
+
+### Encrypt
+
+An example of encrypting to the abcrypt encrypted data format.
+
+### Decrypt
+
+An example of decrypting from the abcrypt encrypted data format.
+
+### Info
+
+An example of reading the Argon2 parameters.
+
+## How to build the example
+
+To build these programs run the following in the project root directory.
+
+```sh
+go build -o . ./...
+```
