+++ examples/decrypt/doc.go
@@ -1,0 +1,23 @@
+// SPDX-FileCopyrightText: 2024 Shun Sakai
+//
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+
+// Decrypt is an example of decrypting a file from the abcrypt encrypted data
+// format.
+//
+// Usage:
+//
+//	decrypt [OPTIONS] <FILE>
+//
+// Arguments:
+//
+//	<FILE>
+//		Input file.
+//
+// Options:
+//
+//	-output <FILE>
+//		Output the result to a file.
+//	-version
+//		Print version number.
+package main
