+++ examples/examples.go
@@ -1,0 +1,9 @@
+// SPDX-FileCopyrightText: 2024 Shun Sakai
+//
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+
+// Package examples contains sample applications for the module.
+package examples
+
+// Version represents the version number of the module.
+const Version = "0.2.0"
