+++ examples/info/doc.go
@@ -1,0 +1,22 @@
+// SPDX-FileCopyrightText: 2024 Shun Sakai
+//
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+
+// Info is an example of reading the Argon2 parameters from a file.
+//
+// Usage:
+//
+//	info [OPTIONS] [FILE]
+//
+// Arguments:
+//
+//	[FILE]
+//		Input file.
+//
+// Options:
+//
+//	-json
+//		Output the encryption parameters as JSON.
+//	-version
+//		Print version number.
+package main
